Changed points possible to double

+++ src/main/java/com/instructure/canvasapi/model/Quiz.java
@@ -36,7 +36,7 @@ public class Quiz extends CanvasModel<Quiz> {
     private QuizPermission permissions;
     private int allowed_attempts;
     private int question_count;
-    private int points_possible;
+    private double points_possible;
     private String due_at;
     private int time_limit;
     private String access_code;
@@ -113,7 +113,7 @@ public class Quiz extends CanvasModel<Quiz> {
         this.question_count = question_count;
     }
 
-    public int getPointsPossible() {
+    public double getPointsPossible() {
         return points_possible;
     }
 
@@ -268,7 +268,7 @@ public class Quiz extends CanvasModel<Quiz> {
         dest.writeParcelable(this.permissions, flags);
         dest.writeInt(this.allowed_attempts);
         dest.writeInt(this.question_count);
-        dest.writeInt(this.points_possible);
+        dest.writeDouble(this.points_possible);
         dest.writeString(this.due_at);
         dest.writeInt(this.time_limit);
         dest.writeString(this.access_code);
@@ -297,7 +297,7 @@ public class Quiz extends CanvasModel<Quiz> {
         this.permissions = in.readParcelable(QuizPermission.class.getClassLoader());
         this.allowed_attempts = in.readInt();
         this.question_count = in.readInt();
-        this.points_possible = in.readInt();
+        this.points_possible = in.readDouble();
         this.due_at = in.readString();
         this.time_limit = in.readInt();
         this.access_code = in.readString();
